+++ Source/WebScheduler.Server/Program.cs
@@ -16,7 +16,6 @@ using Serilog.Formatting.Compact;
 using Orleans.Versions.Compatibility;
 using Orleans.Versions.Selector;
 using WebScheduler.Server.HealthChecks;
-using TaskScheduler.ServiceHost.Server.Interceptors;
 
 public class Program
 {
@@ -147,7 +146,6 @@ public class Program
             })
             .UseIf(RuntimeInformation.IsOSPlatform(OSPlatform.Linux), x => x.UseLinuxEnvironmentStatistics())
             .UseIf(RuntimeInformation.IsOSPlatform(OSPlatform.Windows), x => x.UsePerfCounterEnvironmentStatistics())
-            .AddIncomingGrainCallFilter<TenantValidationInterceptor>()
             .UseDashboard(options => options.BasePath = GetOrleansDashboardOptions(context.Configuration).BasePath);
 
     private static void ConfigureWebHostBuilder(IWebHostBuilder webHostBuilder) =>
